fix(vuln): keep SBOM generation enabled for images with no detectable base OS
AI-assisted per request (ROX-36762): gate SBOM only on missing scan data, not CVE-accuracy notes like OS_UNAVAILABLE, matching CLI/backend behavior.
