+++ ui/apps/platform/src/Containers/Vulnerabilities/components/GenerateSbomModal.tsx
@@ -18,24 +18,6 @@ import useRestMutation from 'hooks/useRestMutation';
 import { generateAndSaveSbom } from 'services/ImageSbomService';
 import type { GenerateSbomImageParams } from 'services/ImageSbomService';
 
-export function getSbomGenerationStatusMessage({
-    isScannerV4Enabled,
-    hasScanMessage,
-}: {
-    isScannerV4Enabled: boolean;
-    hasScanMessage: boolean;
-}): string | undefined {
-    if (!isScannerV4Enabled) {
-        return 'SBOM generation requires Scanner V4';
-    }
-
-    if (hasScanMessage) {
-        return 'SBOM generation is unavailable due to incomplete scan data';
-    }
-
-    return undefined;
-}
-
 export type GenerateSbomModalProps = {
     onClose: () => void;
     image: GenerateSbomImageParams;
